### 2025-08-04 #8724 @editorialbot
[PRE REVIEW]: evoxels: A differentiable physics framework for voxel-based microstructure simulations
**Submitting author:** @daubners (<a href="http://orcid.org/0000-0002-7944-6026">Simon Daubner</a>)
**Repository:** https://github.com/daubners/evoxels
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.1.1
