### 2025-09-24 acc7bec
add mms test function
