+++ evoxels/utils.py
@@ -3,15 +3,17 @@ import sympy as sp
 import sympy.vector as spv
 import evoxels as evo
 from evoxels.problem_definition import SmoothedBoundaryODE
+from evoxels.solvers import TimeDependentSolver
+import contextlib, io
 
 ### Generalized test case
 def rhs_convergence_test(
-    ODE_class,       # an ODE class with callable rhs(field, t)->torch.Tensor (shape [x,y,z])
-    problem_kwargs,  # problem parameters to instantiate ODE
-    test_function,   # exact init_fun(x,y,z)->np.ndarray
-    mask_function=None, 
-    convention="cell_center",
-    dtype="float32",
+    ODE_class,
+    problem_kwargs,
+    test_function,
+    mask_function = None,
+    convention = "cell_center",
+    dtype = "float32",
     powers = np.array([3,4,5,6,7]),
     backend = "torch"
 ):
@@ -22,7 +24,7 @@ def rhs_convergence_test(
     slope arrays have one entry for each provided function.
 
     Args:
-        ODE_class: an ODE class with callable rhs(field, t).
+        ODE_class: ODE class with callable rhs(t, u).
         problem_kwargs: problem-specific parameters to instantiate ODE.
         test_function: single sympy expression or a list of expressions.
         mask_function: static mask for smoothed boundary method.
@@ -38,7 +40,10 @@ def rhs_convergence_test(
             "is not a SmoothedBoundaryODE."
         )
     CS = spv.CoordSys3D('CS')
+    t = sp.symbols('t', real=True)
+
     # Prepare lambdified mask if needed
+    # Assumed to be static i.e. no function of t
     mask = (
         sp.lambdify((CS.x, CS.y, CS.z), mask_function, "numpy")
         if mask_function is not None
@@ -67,25 +72,27 @@ def rhs_convergence_test(
         elif convention == 'staggered_x':
             vf = evo.VoxelFields((2**p + 1, 2**p, 2**p), (1, 1, 1), convention=convention)
         vf.precision = dtype
-        grid = vf.meshgrid()
+        dx[i] = vf.spacing[0]
+
         if backend == 'torch':
             vg = evo.voxelgrid.VoxelGridTorch(vf.grid_info(), precision=vf.precision, device='cpu')
         elif backend == 'jax':
             vg = evo.voxelgrid.VoxelGridJax(vf.grid_info(), precision=vf.precision)
-    
+
+        # Init mask if smoothed boundary ODE
+        numpy_grid = vf.meshgrid()
+        if mask is not None:
+            problem_kwargs["mask"] = mask(*numpy_grid)
+
         # Initialise fields
         u_list = []
         for func in test_functions:
             init_fun = sp.lambdify((CS.x, CS.y, CS.z), func, "numpy")
-            init_data = init_fun(*grid)
+            init_data = init_fun(*numpy_grid)
             u_list.append(vg.init_scalar_field(init_data))
 
         u = vg.concatenate(u_list, 0)
         u = vg.bc.trim_boundary_nodes(u)
-
-        # Init mask if smoothed boundary ODE
-        if mask is not None:
-            problem_kwargs["mask"] = mask(*grid)
 
         ODE = ODE_class(vg, **problem_kwargs)
         rhs_numeric = ODE.rhs(0, u)
@@ -103,14 +110,13 @@ def rhs_convergence_test(
         for j, func in enumerate(test_functions):
             comp = vg.export_scalar_field_to_numpy(rhs_numeric[j:j+1])
             exact_fun = sp.lambdify((CS.x, CS.y, CS.z), rhs_analytic[j], "numpy")
-            exact = exact_fun(*grid)
+            exact = exact_fun(*numpy_grid)
             if convention == "staggered_x":
                 exact = exact[1:-1, :, :]
 
             # Error norm
[…]
