### 2025-09-25 37a75ff
add coupled reaction diffusion MMS example
