### 2026-01-05 #8724 @philipcardiff
Let's continue the review at https://github.com/openjournals/joss-reviews/issues/9733.
