### 2026-01-05 #9733 @arfon
@philipcardiff @daubners – please note that as of today, new submissions entering review will be required to include some additional sections in the paper. You can read more about these requirements here https://joss.readthedocs.io/en/latest/paper.html and the motivations behind this here: https://blog.joss.theoj.org/2026/01/preparing-joss-for-a-generative-ai-future

Any questions, please let me know!
