+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "evoxels"
-version = "0.1.1"
+version = "0.1.2"
 description = "Differentiable physics framework for voxel-based microstructure simulations"
 authors = [
     { name = "Simon Daubner", email = "s.daubner@imperial.ac.uk" }
