### 2026-01-09 daubners/evoxels#36 @david-zwicker
Slightly misleading statement in paper
As part of the review process at https://github.com/openjournals/joss-reviews/issues/9733 I looked at the manuscript and stumbled upon the following statement on Line 77: “no scalable 3D Python implementation exists [for the Cahn-Hilliard equation]”. I believe that this statement is wrong in this generality since there certainly are such implementations – we develop one based on our package https://github.com/zwicker-group/py-pde and I know of at least one other implementation (which is unfortunately not available publicly). I suggest that the respective sentence in the manuscript is adjusted.
