### 2026-01-13 3ee275f
improve install instructions
