### 2026-01-13 538e563
code robustness
