### 2026-01-29 795b375
update paper and references
