+++ paper.md
@@ -38,14 +38,14 @@ Materials science inherently spans disciplines: experimentalists use advanced mi
 
 The differentiable physics framework **evoxels** is based on a fully Pythonic, unified voxel-based approach that integrates segmented 3D microscopy data, physical simulations, inverse modeling, and machine learning.
 
-- At its core is a voxel grid representation compatible with both pytorch and jax to leverage massive parallelization on CPU, GPU and TPU for large microstructures.
+- At its core is a voxel grid representation compatible with both PyTorch and JAX to leverage massive parallelization on CPU, GPU and TPU for large microstructures.
 - Both backends naturally provide high computational performance based on just-in-time compiled kernels and end-to-end gradient-based parameter learning through automatic differentiation.
 - The solver design based on advanced Fourier spectral time-stepping and low-RAM in-place updates enables scaling to hundreds of millions of DOFs on commodity hardware (e.g. forward Cahn-Hilliard simulation with $400^3$ voxels on NVIDIA RTX 500 Ada Laptop GPU with 4GB memory) and billions of DOFs on high end data-center GPUs ($1024^3$ voxels on NVIDIA RTX A6000; more details see \autoref{fig:benchmark}).
 - Its modular design includes comprehensive convergence tests to ensure the right order of convergence and robustness for various combinations of boundary conditions, grid conventions and stencils during rapid prototyping of new PDEs.
 
 While not intended to replace general finite-element or multi-physics platforms, it fills a unique niche for high-resolution voxel workflows, rapid prototyping for structure simulations and materials design, and fully open, reproducible pipelines that bridge imaging, modeling, and data-driven optimization.
 
-From a high-level perspective, evoxels is organized around two core abstractions: VoxelFields and VoxelGrid. VoxelFields provides a uniform, NumPy-based container for any number of 3D fields on the same regular grid, maximizing interoperability with image I/O libraries (e.g. tifffile, h5py, napari, scikit-image) and visualization tools (PyVista, VTK). VoxelGrid couples these fields to either a PyTorch or JAX backend, offering pre-defined boundary conditions, finite difference stencils and FFT libraries as sketched in \autoref{fig:evoxels-code-logic}. The implemented solvers leverage advanced Fourier spectral timesteppers (e.g. semi-implicit, exponential integrators), on-the-fly plotting and integrated wall-time and RAM profiling. A suite of predefined PDE “problems” (e.g. Cahn–Hilliard, reaction-diffusion, multi-phase evolution) can be solved out of the box or extended via user-defined ODEs classes with custom right-hand sides. Integrated convergence tests ensure each discretization achieves the expected order before it ever touches real microscopy data.
+From a high-level perspective, evoxels is organized around two core abstractions: VoxelFields and VoxelGrid. VoxelFields provides a uniform, NumPy-based container for any number of 3D fields on the same regular grid, maximizing interoperability with image I/O libraries (e.g. tifffile, h5py, napari [@napari], scikit-image [@scimage]) and visualization tools (e.g. PyVista [@sullivan2019pyvista] for visualization and VTK export). VoxelGrid couples these fields to either a PyTorch or JAX backend, offering pre-defined boundary conditions, finite difference stencils and FFT libraries as sketched in \autoref{fig:evoxels-code-logic}. The implemented solvers leverage advanced Fourier spectral timesteppers (e.g. semi-implicit, exponential integrators), on-the-fly plotting and integrated wall-time and RAM profiling. A suite of predefined PDE “problems” (e.g. Cahn–Hilliard, reaction-diffusion, multi-phase evolution) can be solved out of the box or extended via user-defined ODEs classes with custom right-hand sides. Integrated convergence tests ensure each discretization achieves the expected order before it ever touches real microscopy data.
 
[…]
