+++ paper.bib
@@ -1,7 +1,6 @@
 @article{Kench2021,
   author = {Kench, Steve and Cooper, Samuel J.},
   doi = {10.1038/s42256-021-00322-1},
-  file = {:home/simond/.local/share/data/Mendeley Ltd./Mendeley Desktop/Downloaded/Kench, Cooper - 2021 - Generating three-dimensional structures from a two-dimensional slice with generative adversarial network-based di.pdf:pdf},
   issn = {2522-5839},
   journal = {Nature Machine Intelligence},
   number = {4},
@@ -42,7 +41,7 @@
 }
 
 @article{Daubner2024_micro,
-    author = {Daubner, S and Nestler, B.},
+    author = {Daubner, S. and Nestler, B.},
     doi = {10.1149/1945-7111/ad9a07},
     issn = {0013-4651},
     journal = {Journal of The Electrochemical Society},
@@ -150,7 +149,7 @@
 }
 
 @article{Daubner2025,
-    author = {Daubner, Simon and Weichel, Marcel and Reder, Martin and Schneider, Daniel and Huang, Qi and Cohen, Alexander E and Bazant, Martin Z. and Nestler, Britta},
+    author = {Daubner, Simon and Weichel, Marcel and Reder, Martin and Schneider, Daniel and Huang, Qi and Cohen, Alexander E. and Bazant, Martin Z. and Nestler, Britta},
     doi = {10.1038/s41524-025-01707-1},
     issn = {2057-3960},
     journal = {npj Computational Materials},
@@ -163,7 +162,7 @@
 }
 
 @article{XinyuLiu2024,
-    author = {{Xinyu Liu}, Xinyu Liu and {Jie Shen}, Jie Shen and {Xiangxiong Zhang}, Xiangxiong Zhang},
+    author = {Liu, Xinyu and Shen, Jie and Zhang, Xiangxiong},
     doi = {10.4208/cicp.OA-2024-0072},
     issn = {1991-7120},
     journal = {Communications in Computational Physics},
@@ -190,4 +189,99 @@
     url = {https://www.nature.com/articles/s41586-023-06393-x},
     volume = {621},
     year = {2023}
+}
+
+@phdthesis{Kidger2021,
+    title={{O}n {N}eural {D}ifferential {E}quations},
+    author={Patrick Kidger},
+    year={2021},
+    school={University of Oxford},
+}
+
+@inproceedings{lienen2022torchode,
+    title = {torchode: A Parallel {ODE} Solver for PyTorch},
+    author = {Marten Lienen and Stephan G{\"u}nnemann},
+    booktitle = {The Symbiosis of Deep Learning and Differential Equations II, NeurIPS},
+    year = {2022},
+    url = {https://openreview.net/forum?id=uiKVKTiUYB0}
+}
+
+@article{sullivan2019pyvista,
+    doi = {10.21105/joss.01450},
+    url = {https://doi.org/10.21105/joss.01450},
+    year = {2019},
+    month = {May},
+    publisher = {The Open Journal},
+    volume = {4},
+    number = {37},
+    pages = {1450},
+    author = {Bane Sullivan and Alexander Kaszynski},
+    title = {{PyVista}: {3D} plotting and mesh analysis through a streamlined interface for the {Visualization Toolkit} ({VTK})},
+    journal = {Journal of Open Source Software}
+}
+
+@misc{napari,
+    author = {napari contributors},
+    doi = {10.5281/zenodo.3555620},
+    year = {2019},
+    title = {{a multi-dimensional image viewer for python}},
+}
+
+@article{scimage,
+    title = {scikit-image: image processing in Python},
[…]
