+++ paper.bib
@@ -149,7 +149,7 @@
 }
 
 @article{Daubner2025,
-    author = {Daubner, Simon and Weichel, Marcel and Reder, Martin and Schneider, Daniel and Huang, Qi and Cohen, Alexander E. and Bazant, Martin Z. and Nestler, Britta},
+    author = {Daubner, S. and Weichel, M. and Reder, M. and Schneider, Daniel and Huang, Qi and Cohen, Alexander E. and Bazant, Martin Z. and Nestler, Britta},
     doi = {10.1038/s41524-025-01707-1},
     issn = {2057-3960},
     journal = {npj Computational Materials},
@@ -221,7 +221,7 @@
 }
 
 @misc{napari,
-    author = {napari contributors},
+    author = {{napari contributors}},
     doi = {10.5281/zenodo.3555620},
     year = {2019},
     title = {{a multi-dimensional image viewer for python}},
