### 2026-01-29 daubners/evoxels#36 @daubners
I changes that repective paragraph of the paper in the actual paper branch @meyer-nils and @david-zwicker
