### 2026-01-29 #9733 @meyer-nils
Sorry for spamming, I hope this will finally compile the current article proof:
