### 2026-01-29 #9733 @daubners
> 👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.09733/joss.09733/10.21105.joss.09733.pdf) 📄 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.09733/joss.09733/10.21105.joss.09733.pdf) 📄 👈

No worries. As far as I can see, this previous compile was already the correct one.
