### 2026-01-29 #9733 @danielskatz
FYI, any command to @editorialbot needs to be the first thing in a new comment.
