### 2026-01-30 #9733 @meyer-nils
@philipcardiff: I have completed my review and checked all items in the checklist. The software meets all JOSS criteria, and the authors have addressed my feedback. I recommend this submission for acceptance.
