### 2026-01-30 #9733 @david-zwicker
I'm also happy to report I completed my report and recommend acceptance. The authors addressed all issues I could identify and provide a nice piece of software with a useful description in form of the paper!
