+++ evoxels/solvers.py
@@ -4,6 +4,9 @@ from typing import Callable, Any, Type
 from abc import ABC, abstractmethod
 from timeit import default_timer as timer
 import sys
+import warnings
+import numpy as np
+import matplotlib.pyplot as plt
 from .problem_definition import ODE
 from .timesteppers import TimeStepper
 
@@ -111,15 +114,49 @@ class BaseSolver(ABC):
         if verbose:
             self.profiler.print_memory_stats(start, end, max_iters)
 
-    def _handle_outputs(self, u, frame, time, slice_idx, vtk_out, verbose, plot_bounds, colormap):
+    def _export_fields(self, u_out):
+        for i, name in enumerate(self.fieldnames):
+            self.vf.set_field(name, self.vg.export_scalar_field_to_numpy(u_out[i:i+1]))
+
+    def _plot_frame(self, fields, time, colormap, plot_bounds):
+        """Plot a single frame of the specified field."""
+        slice_idx = self.vf.Nz // 2
+        slice = self.vg.to_numpy(fields[0, :, :, slice_idx])
+
+        if not hasattr(self, "_fig"):
+            self._fig, self._ax0 = plt.subplots(nrows=1, ncols=1)
+            start1 = self.vf.origin[0] - self.vf.spacing[0] / 2
+            start2 = self.vf.origin[1] - self.vf.spacing[1] / 2
+            end1 = self.vf.domain_size[0] - start1
+            end2 = self.vf.domain_size[1] - start2
+            extent = [start1, end1, start2, end2]
+
+            if plot_bounds is not None:
+                self._im0 = self._ax0.imshow(slice.T, cmap=colormap, origin='lower', extent=extent,\
+                                             vmin=plot_bounds[0], vmax=plot_bounds[1])
+            else:
+                self._im0 = self._ax0.imshow(slice.T, cmap=colormap, origin='lower', extent=extent)
+            ratio = np.clip((end2 - start2) / (end1 - start1), 0, 1)
+            self._cbar = self._fig.colorbar(self._im0, ax=self._ax0, shrink=ratio)
+            self._ax0.set_xlabel("X")
+            self._ax0.set_ylabel("Y")
+        else:
+            self._im0.set_data(slice.T)
+            if plot_bounds is None:
+                self._im0.set_clim(float(slice.min()), float(slice.max()))
+                self._cbar.update_normal(self._im0)
+
+        self._ax0.set_title(f"Slice {slice_idx} of {self.fieldnames[0]} in z at time {time}")
+        self._fig.canvas.draw_idle()
+        self._fig.canvas.flush_events()
+
+    def _handle_outputs(self, u, frame, time, vtk_out, verbose, plot_bounds, colormap):
         """Store results and optionally plot or write them to disk."""
         if getattr(self, 'problem', None) is not None:
             u_out = self.vg.bc.trim_ghost_nodes(self.problem.pad_bc(u))
         else:
             u_out = self.vg.bc.trim_ghost_nodes(self.vg.pad_zeros(u))
-
-        for i, name in enumerate(self.fieldnames):
-            self.vf.set_field(name, self.vg.export_scalar_field_to_numpy(u_out[i:i+1]))
+        self._export_fields(u_out)
 
         if verbose:
             self.profiler.update_memory_stats()
@@ -134,8 +171,21 @@ class BaseSolver(ABC):
             self.vf.export_to_vtk(filename=filename, field_names=self.fieldnames)
 
         if verbose == 'plot':
-            clear_output(wait=True)
-            self.vf.plot_slice(self.fieldnames[0], slice_idx, time=time, colormap=colormap, value_bounds=plot_bounds)
+            if not plt.isinteractive():
+                plt.ion()
+
[…]
