+++ paper.md
@@ -12,6 +12,7 @@ authors:
     corresponding: true
     affiliation: 1
   - name: Alexander E. Cohen
+    orcid: 0000-0002-5284-6775
     affiliation: 2
   - name: Benjamin Dörich
     orcid: 0000-0001-5840-2270
