+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "evoxels"
-version = "0.1.2"
+version = "1.0.0"
 description = "Differentiable physics framework for voxel-based microstructure simulations"
 authors = [
     { name = "Simon Daubner", email = "s.daubner@imperial.ac.uk" }
