### 2026-02-04 #9733 @daubners
@philipcardiff , I've checked the authors, affiliations and ORCIDS. Updated the version and created a new release which is updated both on pypi and now available on zenodo under doi [10.5281/zenodo.18489155](https://doi.org/10.5281/zenodo.18489155)
