### 2026-02-18 f940e30
Punctuation and formatting in paper.md
Punctuation and formatting changes throughout the document for improved readability and consistency.

This commit relates to the JOSS review at https://github.com/openjournals/joss-reviews/issues/9733#issuecomment-3908196608.
