### 2026-02-18 #9733 @philipcardiff
HI @daubners, I have proposed a few minor edits to the paper at https://github.com/daubners/evoxels/pull/43.
