### 2026-02-20 #9733 @philipcardiff
@Kevin-Mattheus-Moerman, this submission is ready for processing, but I am getting an error when running `@editorialbot recommend-accept`, which I cannot decipher.
