### 2026-02-25 #9733 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
