### 2026-03-15 #9733 @Kevin-Mattheus-Moerman
@daubners As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Apologies for the delay in processing these last steps. Below are some final checks some of which require your attention:

- [x] Copyedit pass/checked for typos
👉 Both American English (e.g. `optimization` and `organization`) and British English are used (e.g. `Visualisation` and `organising`) please choose one form and use it consistently.
