+++ paper.md
@@ -48,7 +48,7 @@ While not intended to replace general finite-element or multi-physics platforms,
 
 At a high level, evoxels is organized around two core abstractions: VoxelFields and VoxelGrid. VoxelFields provides a uniform, NumPy-based container for any number of 3D fields on the same regular grid, maximizing interoperability with image I/O libraries (e.g., tifffile, h5py, napari [@napari], scikit-image [@scimage]) and visualization tools (e.g., PyVista [@sullivan2019pyvista] for visualization and VTK export). VoxelGrid couples these fields to either a PyTorch or JAX backend, offering pre-defined boundary conditions, finite difference stencils, and FFT libraries as sketched in \autoref{fig:evoxels-code-logic}. The implemented solvers leverage advanced Fourier spectral timesteppers (e.g., semi-implicit, exponential integrators), on-the-fly plotting, and integrated wall-time and RAM profiling. A suite of predefined PDE “problems” (e.g., Cahn–Hilliard, reaction-diffusion, multi-phase evolution) can be solved out of the box or extended via user-defined ODE classes with custom right-hand sides. Integrated convergence tests ensure that each discretization achieves the expected order before it is applied to real microscopy data.
 
-![Visualisation of package concept. The VoxelFields class acts as the user interface for organising 3D fields on a regular grid, including plotting and export functions. Solvers are assembled in a modular fashion. The chosen timestepper and ODE class are just-in-time compiled (green becomes a single kernel) based on the specified VoxelGrid backend.\label{fig:evoxels-code-logic}](evoxels-code-logic.png)
+![Visualization of package concept. The VoxelFields class acts as the user interface for organizing 3D fields on a regular grid, including plotting and export functions. Solvers are assembled in a modular fashion. The chosen timestepper and ODE class are just-in-time compiled (green becomes a single kernel) based on the specified VoxelGrid backend.\label{fig:evoxels-code-logic}](evoxels-code-logic.png)
 
 **evoxels** is aimed squarely at researchers who need a “plug-in-your-image, get-your-answer” workflow for digital materials science and inverse design. Experimentalists can feed segmented FIB-SEM or X-ray tomograms directly into high-performance simulations; computational scientists and modelers benefit from a truly open, reproducible framework. It speaks to anyone who wants special-purpose solvers for representative volume elements - without the overhead of mesh generation - while still offering the flexibility to develop new solvers, test boundary conditions, and incorporate machine-learning-driven optimization. evoxels provides both the turnkey usability of a specialized package and the extensibility of a low-level research toolkit, e.g., benchmarking tortuosity, fitting diffusion coefficients, or prototyping novel phase-field models.
 
